+++ pom.xml
@@ -2,7 +2,7 @@
     xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
     <modelVersion>4.0.0</modelVersion>
 
-    <groupId>com.example</groupId>
+    <groupId>com.ahumadamob</groupId>
     <artifactId>catalogo-back</artifactId>
     <version>0.0.1-SNAPSHOT</version>
     <packaging>jar</packaging>
