+++ src/main/java/com/ahumadamob/catalogo/CatalogoBackApplication.java
@@ -1,0 +1,12 @@
+package com.ahumadamob.catalogo;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CatalogoBackApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(CatalogoBackApplication.class, args);
+    }
+}
