+++ src/main/java/com/ahumadamob/catalogo/controller/CategoryController.java
@@ -1,0 +1,51 @@
+package com.ahumadamob.catalogo.controller;
+
+import com.ahumadamob.catalogo.dto.CategoriaDto;
+import com.ahumadamob.catalogo.service.CategoryService;
+import jakarta.validation.Valid;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
+
+@RestController
+@RequestMapping("/api/v1/categorias")
+@RequiredArgsConstructor
+public class CategoryController {
+
+    private final CategoryService service;
+
+    @GetMapping("/")
+    public ResponseEntity<List<CategoriaDto>> getAllCategories() {
+        return ResponseEntity.ok(service.getAllCategories());
+    }
+
+    @GetMapping("/{id}")
+    public ResponseEntity<CategoriaDto> getCategoryById(@PathVariable Long id) {
+        return ResponseEntity.ok(service.getCategoryById(id));
+    }
+
+    @PostMapping("/")
+    public ResponseEntity<CategoriaDto> createCategory(@Valid @RequestBody CategoriaDto dto) {
+        CategoriaDto created = service.createCategory(dto);
+        return ResponseEntity.status(HttpStatus.CREATED).body(created);
+    }
+
+    @PutMapping("/{id}")
+    public ResponseEntity<CategoriaDto> updateCategory(@PathVariable Long id, @Valid @RequestBody CategoriaDto dto) {
+        return ResponseEntity.ok(service.updateCategory(id, dto));
+    }
+
+    @DeleteMapping("/{id}")
+    public ResponseEntity<Void> deleteCategory(@PathVariable Long id) {
+        service.deleteCategory(id);
+        return ResponseEntity.noContent().build();
+    }
+
+    @GetMapping("/arbol")
+    public ResponseEntity<List<CategoriaDto>> getCategoryTree() {
+        return ResponseEntity.ok(service.getCategoryTree());
+    }
+}
