+++ src/main/java/com/ahumadamob/catalogo/dto/ItemDto.java
@@ -1,0 +1,16 @@
+package com.ahumadamob.catalogo.dto;
+
+import jakarta.validation.constraints.NotNull;
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class ItemDto {
+    @NotNull(message = "el nombre es obligatorio")
+    private String nombre;
+}
