+++ src/main/java/com/ahumadamob/catalogo/dto/Link.java
@@ -1,0 +1,15 @@
+package com.ahumadamob.catalogo.dto;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+@Data
+@Builder
+@NoArgsConstructor
+@AllArgsConstructor
+public class Link {
+    private String rel;
+    private String href;
+}
