+++ src/main/java/com/ahumadamob/catalogo/exception/ResourceNotFoundException.java
@@ -1,0 +1,7 @@
+package com.ahumadamob.catalogo.exception;
+
+public class ResourceNotFoundException extends RuntimeException {
+    public ResourceNotFoundException(String message) {
+        super(message);
+    }
+}
