+++ src/main/java/com/ahumadamob/catalogo/repository/CategoriaRepository.java
@@ -1,0 +1,7 @@
+package com.ahumadamob.catalogo.repository;
+
+import com.ahumadamob.catalogo.entity.Categoria;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
+}
