--- src/main/java/com/example/catalogo/CatalogoBackApplication.java
@@ -1,12 +1,0 @@
-package com.example.catalogo;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class CatalogoBackApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(CatalogoBackApplication.class, args);
-    }
-}
