--- src/main/java/com/example/catalogo/dto/ItemDto.java
@@ -1,16 +1,0 @@
-package com.example.catalogo.dto;
-
-import jakarta.validation.constraints.NotNull;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class ItemDto {
-    @NotNull(message = "el nombre es obligatorio")
-    private String nombre;
-}
