--- src/main/java/com/example/catalogo/dto/Link.java
@@ -1,15 +1,0 @@
-package com.example.catalogo.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Link {
-    private String rel;
-    private String href;
-}
