--- src/main/java/com/example/catalogo/dto/Pagination.java
@@ -1,17 +1,0 @@
-package com.example.catalogo.dto;
-
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
-import lombok.NoArgsConstructor;
-
-@Data
-@Builder
-@NoArgsConstructor
-@AllArgsConstructor
-public class Pagination {
-    private int page;
-    private int pageSize;
-    private long totalElements;
-    private int totalPages;
-}
