--- src/main/java/com/example/catalogo/repository/ItemRepository.java
@@ -1,7 +1,0 @@
-package com.example.catalogo.repository;
-
-import com.example.catalogo.entity.Item;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface ItemRepository extends JpaRepository<Item, Long> {
-}
