--- src/main/java/com/miempresa/catalogo/exception/ResourceNotFoundException.java
@@ -1,7 +1,0 @@
-package com.miempresa.catalogo.exception;
-
-public class ResourceNotFoundException extends RuntimeException {
-    public ResourceNotFoundException(String message) {
-        super(message);
-    }
-}
