--- src/main/java/com/miempresa/catalogo/repository/CategoriaRepository.java
@@ -1,7 +1,0 @@
-package com.miempresa.catalogo.repository;
-
-import com.miempresa.catalogo.entity.Categoria;
-import org.springframework.data.jpa.repository.JpaRepository;
-
-public interface CategoriaRepository extends JpaRepository<Categoria, Long> {
-}
